fix(frontend): move viewport to useWorkflowViewport
